Switched from LWP::UserAgent perl module to wget
LWP::UserAgent module is a part of a rather heavy LWP suite. We don't
really need the full blown functionality of LWP for a simple HTTP
GET request with a User-Agent string.  So, switching to a wget
binary, which is more often installed, and even if not, is much
smaller than LWP suite.

Hence, removing the perl module-related section from the Known issues
in README.