Remove obsolete identity setting in .bashrc
This was used before for generating a correct .gitconfig with
Ansible template.  But now that the file can be edited directly,
since its outside of the role, there is no more need for setting
the identity variables.

Also, hardcoding the values of my own identity in the .bashrc makes
it much less reusable.