Converted conky config to v1.10+ format
The new version of conky comes with it's own and also much
different configuration file format.  Luckily, there is a
conversion script that comes with the new version.

On Fedora 33, it is located in `/usr/share/doc/conky/convert.lua`.

It does the job perfectly.  However, i3wm needs a full blown restart
after this.